processing/requirements.txt
Flask~=2.1
requests==2.23.0
redis==3.4.1
rq==1.2.2
ffmpeg-python==0.2.0
boto3==1.28.68

